Initial commit, headstart on libraries, makefile and environment
added milana's libft, and my own printf into the library directory.
tinkered around with the makefiles until they worked, NOTE: had to
modify some of the cflags of the libft makefile, as the symbols weren't
getting indexed into the archive properly, idk which one specifically
did it, but it was definitely one of the one i commented out.
implemented the environment, and extremely basic error handling (i.e,
error messages).